go mod and vendor update (#3534)
* go mod and vendor update

* unit test upd

* go mod tidy, go mod vendor on mac

* try to fix windows winio dependency change in docker lib

* hack: fix windows build

* upgrade containernetworking/plugins

* resolve some accidental downgrades

* fix github actions

+++ agent/handlers/task_server_setup.go
@@ -78,7 +78,7 @@ func taskServerSetup(credentialsManager credentials.Manager,
 
 	agentAPIV1HandlersSetup(muxRouter, state, credentialsManager, cluster, region, apiEndpoint, acceptInsecureCert)
 
-	limiter := tollbooth.NewLimiter(int64(steadyStateRate), nil)
+	limiter := tollbooth.NewLimiter(float64(steadyStateRate), nil)
 	limiter.SetOnLimitReached(handlersutils.LimitReachedHandler(auditLogger))
 	limiter.SetBurst(burstRate)
